fix: 에러 함수 main.py에 분리
