+++ config-server/error.py
@@ -1,0 +1,17 @@
+def infra_error(step, error, detail, rollback=None, **extra):
+    body = {
+        "step": step,
+        "error": error,
+        "detail": detail,
+        "rollback": rollback or {},
+    }
+    body.update({key: value for key, value in extra.items() if value is not None})
+    return body
+
+
+def k8s_error_fields(exc):
+    return {
+        "k8s_status": exc.status,
+        "k8s_reason": exc.reason,
+        "k8s_body": exc.body,
+    }
